backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
drf-yasg==1.21.7
python-dotenv==1.0.1
django-cors-headers==4.3.1
Pillow==10.2.0
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.1
couchdb==1.2
python-cloudant==2.16.0 